fix(ui): address pr review comments on copy fixes
- Move SupabaseUser and SupabaseUserEmail interfaces to shared package
  (packages/shared/src/interfaces/supabase.interface.ts) per dev rules
- Import interfaces from @lfx-one/shared/interfaces in supabase.service.ts
- Downgrade stub log calls from logger.warning to logger.debug — stub
  behaviour is expected and intentional, not an anomaly
