+++ apps/lfx-one/src/server/services/supabase.service.ts
@@ -1,22 +1,8 @@
 // Copyright The Linux Foundation and each contributor to LFX.
 // SPDX-License-Identifier: MIT
 
+import { SupabaseUser, SupabaseUserEmail } from '@lfx-one/shared/interfaces';
 import { logger } from './logger.service';
-
-/** Minimal Supabase user shape used by badge fallback flows. */
-export interface SupabaseUser {
-  id: string;
-  username: string;
-}
-
-/**
- * Supabase email record used by badge email resolution.
- * Fields use snake_case to match the raw Supabase wire format.
- */
-export interface SupabaseUserEmail {
-  email: string;
-  is_verified: boolean;
-}
 
 // TODO: Replace this stub with a real Supabase integration once the service is implemented.
 /** Temporary Supabase service stub — returns null/empty so callers fall back to OIDC email. */
@@ -27,7 +13,7 @@ export class SupabaseService {
    * @returns The user record, or null if not found. Stub always returns null.
    */
   public async getUser(username: string): Promise<SupabaseUser | null> {
-    logger.warning(undefined, 'supabase_get_user', 'SupabaseService is a stub — returning null', { username });
+    logger.debug(undefined, 'supabase_get_user', 'SupabaseService is a stub — returning null', { username });
     return null;
   }
 
@@ -37,7 +23,7 @@ export class SupabaseService {
    * @returns List of email records. Stub always returns an empty list.
    */
   public async getUserEmails(userId: string): Promise<SupabaseUserEmail[]> {
-    logger.warning(undefined, 'supabase_get_user_emails', 'SupabaseService is a stub — returning empty list', { userId });
+    logger.debug(undefined, 'supabase_get_user_emails', 'SupabaseService is a stub — returning empty list', { userId });
     return [];
   }
 }
