+++ packages/shared/src/interfaces/index.ts
@@ -134,3 +134,6 @@ export * from './multi-persona-dashboard.interface';
 
 // Transaction interfaces
 export * from './transaction.interface';
+
+// Supabase interfaces
+export * from './supabase.interface';
